% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\name{logitMod}
\alias{logitMod}
\title{Alternative Logistic Regression Implementation}
\usage{
logitMod(formula, data, precision = 1e-10, iterMax = 25)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{an optional data frame or coercible to one. If empty, the
variables are taken from the Global Environment.}

\item{precision}{degree of precision of optimisation algorithm.}

\item{iterMax}{maximum number of iterations of optimisation algorithm.}
}
\value{
\code{logitMod} returns an object of \code{\link[base]{class}}
    \emph{logitMod}.
}
\description{
\code{logitMod} computes a logistic regression for a binary response variable
   via \code{\link{newtonRaphson}} estimation method.
}
\examples{
set.seed(42)
X1 <- rnorm(100);    X2 <- rnorm(100);
points <- 2 * X1 - 3 * X2
y <- rbinom(100, 1, exp(points) / (1 + exp(points)))
X <- cbind(b1 = 1, b2 = X1, b3 = X2)
fit <- logitMod(y ~ X1 + X2)
}
