% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\name{newtonRaphson}
\alias{newtonRaphson}
\title{Newton-Raphson for Logistic Regression Estimation}
\usage{
newtonRaphson(y, X, precision = 1e-07, iterMax = 25)
}
\arguments{
\item{y}{Matrix/vector containing the response variable.}

\item{X}{Matrix/vector containing the design matrix (including intercept).}

\item{precision}{Degree of precision of algorithm.}

\item{iterMax}{Maximum number of iterations of optimisation algorithm.}
}
\value{
A list with maximum likelihood estimation results for further
  computation.
}
\description{
This function calculates the maximum likelihood for binary logistic
regression via the Newton Raphson algorithm.
}
\details{
This algorithm is used on the \code{\link{logitModel}} regression estimation
function.
}
\examples{
set.seed(42)
X1 <- rnorm(100);    X2 <- rnorm(100);
points <- 2 * X1 - 3 * X2
y <- rbinom(100, 1, exp(points) / (1 + exp(points)))
X <- cbind(b1 = 1, b2 = X1, b3 = X2)
fit <- newtonRaphson(y = y, X = X)

}
\references{
Agresti, A (2013) Categorical Data Analysis,
John Wiley & Sons, 2013, Volume 792 of Wiley Series in Probability
and Statistics, ISSN 1940-6517

Czepiel, S.A. (2002) Maximum Likelihood Estimation of Logistic
Regression Models: Theory and Implementation. Available at
\href{https://czep.net/stat/mlelr.pdf}{czep.net/stat/mlelr.pdf}. [visited:
25.03.2018]
}
