% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\name{pairs.logitModel}
\alias{pairs.logitModel}
\title{S3 Pairs Method for logitModel Objects}
\usage{
\method{pairs}{logitModel}(model, nRow, single = FALSE, col, ask, ...)
}
\arguments{
\item{model}{an object of class logitModel.}

\item{nRow}{optional numeric vector indicating the number of rows the overall
plot should have.}

\item{single}{boolean to indicate whether each variable should be plotted
against the response variable individually.}
}
\description{
This internal function defines a \code{\link{pairs}} method for an object
of logitModel class. The main idea is to plot all explanatory variables
agaist the response variable of the model in order to quickly assess the
influence of each one.
}
\examples{
set.seed(42)
n <- 100
dummies <- sample(LETTERS[1:5], n, replace = TRUE)
X1 <- rnorm(n); X2 <- rnorm(n)
points <- 2 * X1 - 3 * X2
y <- rbinom(n, 1, exp(points) / (1 + exp(points)))
df <- as.data.frame(cbind(b2 = X1, b3 = X2))
myFitD <- logitModel(y ~ X1 + X2 + dummies)
pairs(myFitD)
pairs(myFitD, single = TRUE)
}
