% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\name{print.logitModel}
\alias{print.logitModel}
\title{S3 Print Method for logitModel Objects}
\usage{
\method{print}{logitModel}(model, ...)
}
\arguments{
\item{model}{a logitModel object}

\item{...}{methods are required to include (at least) the same arguments
as their generic functions.}
}
\value{
a brief summary of the model results.
}
\description{
This internal function defines a \code{\link{print}} method for an object of
logitModel class.
}
