% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\name{print.summary.logitModel}
\alias{print.summary.logitModel}
\title{S3 Print Method for summary.logitModel Objects}
\usage{
\method{print}{summary.logitModel}(model, ...)
}
\arguments{
\item{model}{an object of class "summary.logitModel".}

\item{...}{methods are required to include (at least) the same arguments
as their generic functions.}
}
\value{
a summary of the model results.
}
\description{
This internal function defines a \code{\link{print}} method for an object
of \code{\link{summary.logitModel}} class.
}
