% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\name{summary.logitModel}
\alias{summary.logitModel}
\title{S3 Summary Method for logitModel Objects}
\usage{
\method{summary}{logitModel}(model, ...)
}
\arguments{
\item{model}{a logitModel object}

\item{...}{methods are required to include (at least) the same arguments
as their generic functions.}
}
\value{
a list with more detailed statistics and computations on the
estimated model
}
\description{
This internal function defines a \code{\link{summary}} method for an object
of logitModel class, where also further statistics are computed for model
comparison and for the print.summary method.
}
