% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModell.R
\name{logitMod}
\alias{logitMod}
\title{Alternative Interface for Logistic Regression Modelling}
\usage{
logitMod(formula, data, precision = 1e-10, iterMax = 100)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{a data frame or coercible to one}

\item{precision}{degree of precision of optimisation algorithm}

\item{iterMax}{maximum number of iterations of optimisation algorithm}
}
\value{
\code{logitMod} returns an object of \code{\link[base]{class}}
    \emph{logitMod}.
}
\description{
\code{logitMod} computes a logistic regression for a binary response variable
   via \code{\link{MLE}} estimation method.
}
\examples{

}
